requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
colour==0.1.4
idna==2.5
nanoleaf==0.4.1
python-dotenv==0.6.4
requests==2.18.1
urllib3==1.21.1
